+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [1.1.1] 2022-05-16
+
+### Fixed
+
+- DatasetParameterType provides labels now
+- DatasetParameterType returns combined dataset ID now
+
 ## [1.1.0] 2022-05-04
 
 ### Added